[Dockerfile] HACK: add dependencies that are on master but not on dev.
We don't know why we enable the extensions. But they fail due to
missing modules. It turns out something obscurely changed between
master and dev, and we tracked it down to a change in the
some dependency in kso_utils. Until the reason for that
has been figured out, let's just add it manually.
There is an issue created for this, see issue #225 in kso.