[build-container.yml] fit in the CI-pipeline
The CI-pipeline will consist of 3 steps, of which build-container is the
first one. The others will be: NB-tests and update-container. To make it
fit in the entire pipeline:

* Change the 'on' settings to only build when docker/requirement changed

* The tag of the image will be the sha code. This tag will be changed into
dev or master if the NB-tests pass and the workflow was triggered by a
push. This will be done in another workflow: 'update-container.yml'