[NB-widget-tests.yml] workflow with tests
Run the tests in a container that is already on the GHCR. Based on if
the workflow build-container.yml has ran or not, the image it will load
has the tag dev, master or the sha-code.

This workflow contains 3 jobs:

1: Configure the gitlub runner. Without it we do not have permissions to check out the github repository from within our container. Therefore this extra job is needed.

2. Determine the image tag as described above

3. The job that runs the tests. For this we install pytest
(MAYBE THIS SHOULD BECOME ONE OF THE REQUIREMENTS?) and run the tests
with python3 -m pytest ...file...

During the tests, the WandB directories are set to a temp directory,
instead of using the directories of the code itself.